refactor: remove github dependencies from ShreksWebhookEventProcessor
