+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Applicaiton/GithubEventNotifierExtensions.cs
@@ -1,3 +1,4 @@
+using Kysect.Shreks.Application.Dto.Study;
 using Kysect.Shreks.Common.Exceptions;
 using Microsoft.Extensions.Logging;
 
@@ -22,6 +23,30 @@ public static class GithubEventNotifierExtensions
         catch (Exception e)
         {
             logger.LogError(e, "Failed to send exception message to user pull request.");
+        }
+    }
+
+    public static async Task NotifySubmissionUpdate(
+        this IPullRequestCommitEventNotifier pullRequestCommitEventNotifier,
+        SubmissionDto submission,
+        ILogger logger,
+        bool sendComment = false,
+        bool sendCommitComment = false)
+    {
+        string message = $"Submission {submission.Code} was updated." +
+                         $"\nState: {submission.State}" +
+                         $"\nDate: {submission.SubmissionDate}";
+
+        if (sendComment)
+        {
+            logger.LogInformation("Notify comment posted into PR: " + message);
+            await pullRequestCommitEventNotifier.SendCommentToPullRequest(message);
+        }
+
+        if (sendCommitComment)
+        {
+            logger.LogInformation("Notify posted as commit comment: " + message);
+            await pullRequestCommitEventNotifier.SendCommentToTriggeredCommit(message);
         }
     }
 }
