+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Applicaiton/IShreksWebhookEventProcessor.cs
@@ -1,0 +1,15 @@
+using Kysect.Shreks.Application.Dto.Github;
+using Microsoft.Extensions.Logging;
+
+namespace Kysect.Shreks.Integration.Github.Applicaiton;
+
+public interface IShreksWebhookEventProcessor
+{
+    Task ProcessPullRequestReopenWebhookAsync(GithubPullRequestDescriptor pullRequestDescriptor, ILogger repositoryLogger, IPullRequestCommitEventNotifier pullRequestCommitEventNotifier, bool? isMerged);
+    Task ProcessPullRequestUpdateWebhookAsync(GithubPullRequestDescriptor pullRequestDescriptor, ILogger repositoryLogger, IPullRequestCommitEventNotifier pullRequestCommitEventNotifier, CancellationToken cancellationToken);
+    Task ProcessPullRequestClosedWebhookAsync(GithubPullRequestDescriptor pullRequestDescriptor, ILogger repositoryLogger, IPullRequestCommitEventNotifier pullRequestCommitEventNotifier, bool merged);
+    Task ProcessPullRequestReviewCommentWebhookAsync(GithubPullRequestDescriptor pullRequestDescriptor, ILogger repositoryLogger, IPullRequetsEventNotifier pullRequestEventNotifier, string? comment);
+    Task ProcessPullRequestReviewRequestChangesWebhookAsync(GithubPullRequestDescriptor pullRequestDescriptor, ILogger repositoryLogger, IPullRequetsEventNotifier pullRequestEventNotifier, string? reviewBody);
+    Task ProcessPullRequestReviewApproveWebhookAsync(GithubPullRequestDescriptor pullRequestDescriptor, ILogger repositoryLogger, IPullRequetsEventNotifier pullRequestEventNotifier, string? commentBody);
+    Task ProcessIssueCommentCreateWebhookAsync(GithubPullRequestDescriptor pullRequestDescriptor, ILogger repositoryLogger, IPullRequestCommentEventNotifier pullRequestCommentEventNotifier, string issueCommentBody);
+}
